# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Pickup Sports App API
  version: 1.0.0
  description: |
    REST API for Pickup Sports App (Java 17, Spring Boot). Includes authentication, game management, search, notifications, profiles, venues, Nepal market features, media presign, and push subscriptions. JWT Bearer auth used for protected endpoints. Some features may be gated by feature flags or env properties.
servers:
  - url: http://localhost:8080
    description: Local development
security:
  - bearerAuth: []
...
paths:
...
  /profiles/me:
    get:
      tags: [Profiles]
      summary: Get my profile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Profile
          headers:
            ETag:
              schema: { type: string }
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfileDTO'
    put:
      tags: [Profiles]
      summary: Update my profile
      security:
        - bearerAuth: []
      parameters:
        - in: header
          name: If-Match
          schema:
            type: string
        - in: header
          name: If-Unmodified-Since
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserProfileRequest'
      responses:
        '200':
          description: Updated
          headers:
            ETag:
              schema: { type: string }
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfileDTO'
  /profiles/{id}:
    get:
      tags: [Profiles]
      summary: Get a user profile by id
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '200':
          description: Profile
          headers:
            ETag:
              schema: { type: string }
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfileDTO'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UpdateUserProfileRequest:
      type: object
      properties:
        bio: { type: string }
        avatarUrl: { type: string }
        skillLevel: { type: string }
        age: { type: integer }
        position: { type: string }
        contactPreference: { type: string }
    UserProfileDTO:
      type: object
      properties:
        id: { type: integer, format: int64 }
        username: { type: string }
        bio: { type: string }
        avatarUrl: { type: string }
        skillLevel: { type: string }
        age: { type: integer }
        position: { type: string }
        contactPreference: { type: string }
...